backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
pydantic
python-dotenv
mysql-connector-python
alembic
pymysql==1.1.2
cryptography
httpx
pydantic-settings>=2.0.0
python-jose
python-dotenv==1.1.1
celery==5.5.3
redis==5.0.8
requests==2.32.4
openai==1.97.1
pillow==11.3.0
pydantic[email]
python-multipart
